chore: move dashboard metrics service to dashboard service
